chore: remove starknet-contract-class (#766)
* Remove `override: true` from dtolnay actions

It's not a valid option

* Move starknet-contract-class to SiR

* Remove `ParsedContractClass`
