--- crates/starknet-contract-class/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "starknet-contract-class"
-version = "0.2.0"
-edition = "2021"
-description = "Crate that parses a Starknet contract class from API"
-license = "Apache-2.0"
-
-[dependencies]
-serde = "1.0.156"
-serde_json = "1.0.94"
-getset = "0.1.2"
-cairo-vm = { workspace=true, features = ["cairo-1-hints"]}
-starknet_api = { workspace = true }
