+++ src/transaction/deploy.rs
@@ -1,3 +1,4 @@
+use crate::services::api::contract_classes::deprecated_contract_class::EntryPointType;
 use crate::{
     core::{
         contract_address::compute_deprecated_class_hash, errors::state_errors::StateError,
@@ -26,7 +27,6 @@ use crate::{
 };
 use cairo_vm::felt::Felt252;
 use num_traits::Zero;
-use starknet_contract_class::EntryPointType;
 
 use super::Transaction;
