+++ tests/complex_contracts/amm_contracts/amm.rs
@@ -4,8 +4,8 @@ use cairo_vm::vm::runners::builtin_runner::HASH_BUILTIN_NAME;
 use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 use cairo_vm::{felt::Felt252, vm::runners::builtin_runner::RANGE_CHECK_BUILTIN_NAME};
 use num_traits::Zero;
-use starknet_contract_class::EntryPointType;
 use starknet_in_rust::definitions::block_context::BlockContext;
+use starknet_in_rust::EntryPointType;
 use starknet_in_rust::{
     execution::{CallInfo, CallType},
     services::api::contract_classes::deprecated_contract_class::ContractClass,
