+++ tests/complex_contracts/amm_contracts/amm_proxy.rs
@@ -1,10 +1,10 @@
 use crate::complex_contracts::utils::*;
 use cairo_vm::felt::Felt252;
 use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
-use starknet_contract_class::EntryPointType;
 use starknet_crypto::FieldElement;
 use starknet_in_rust::definitions::block_context::BlockContext;
 use starknet_in_rust::services::api::contract_classes::deprecated_contract_class::ContractClass;
+use starknet_in_rust::EntryPointType;
 use starknet_in_rust::{
     execution::{CallInfo, CallType},
     state::{cached_state::CachedState, state_api::StateReader},
