+++ tests/delegate_call.rs
@@ -4,7 +4,7 @@ mod cairo_1_syscalls;
 
 use cairo_vm::felt::Felt252;
 use num_traits::{One, Zero};
-use starknet_contract_class::EntryPointType;
+use starknet_in_rust::EntryPointType;
 use starknet_in_rust::{
     definitions::{block_context::BlockContext, constants::TRANSACTION_VERSION},
     execution::{
