+++ tests/fibonacci.rs
@@ -5,8 +5,8 @@ use cairo_lang_starknet::casm_contract_class::CasmContractClass;
 use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 use cairo_vm::{felt::Felt252, vm::runners::builtin_runner::RANGE_CHECK_BUILTIN_NAME};
 use num_traits::Zero;
-use starknet_contract_class::EntryPointType;
 use starknet_in_rust::definitions::block_context::BlockContext;
+use starknet_in_rust::EntryPointType;
 use starknet_in_rust::{
     definitions::constants::TRANSACTION_VERSION,
     execution::{
