+++ tests/internals.rs
@@ -12,7 +12,6 @@ use cairo_vm::vm::{
 use lazy_static::lazy_static;
 use num_bigint::BigUint;
 use num_traits::{Num, One, ToPrimitive, Zero};
-use starknet_contract_class::EntryPointType;
 use starknet_in_rust::core::errors::state_errors::StateError;
 use starknet_in_rust::definitions::constants::{
     DEFAULT_CAIRO_RESOURCE_FEE_WEIGHTS, VALIDATE_ENTRY_POINT_SELECTOR,
@@ -23,6 +22,7 @@ use starknet_in_rust::services::api::contract_classes::deprecated_contract_class
 use starknet_in_rust::state::ExecutionResourcesManager;
 use starknet_in_rust::transaction::{DeclareV2, Deploy};
 use starknet_in_rust::CasmContractClass;
+use starknet_in_rust::EntryPointType;
 use starknet_in_rust::{
     definitions::{
         block_context::{BlockContext, StarknetChainId, StarknetOsConfig},
